ARM: Allow this header to be included by assembly files
The constants defined in this file are equally useful in assembly and C
source files. The arm64 architecture version of this file allows
inclusion in both assembly and C source files, so this this commit adds
that capability to the arm architecture version so that the constants
don't need to be defined in multiple places.

Signed-off-by: Doug Berger <opendmb@gmail.com>
Signed-off-by: Florian Fainelli <f.fainelli@gmail.com>
Tested-by: Mylène Josserand <mylene.josserand@bootlin.com>
Signed-off-by: Florian Fainelli <f.fainelli@gmail.com>